+++ space-invaders/core/src/com/samsung/business/spaceinvaders/entity/PlayerShot.java
@@ -1,19 +1,21 @@
-package com.samsung.business.spaceinvaders.byty;
+package com.samsung.business.spaceinvaders.entity;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Rectangle;
-import com.samsung.business.spaceinvaders.zarzadzanie.GraphicsManager;
+import com.samsung.business.spaceinvaders.manager.GraphicsManager;
 
 /**
- * Created by lb_lb on 01.11.17.
+ * Created by lb_lb on 05.11.17.
  */
-public class EnemyShot implements Shot {
+public class PlayerShot implements Shot {
+    private static final int HEIGHT = 480;
+
     private Rectangle rectangle;
     private GraphicsManager.Graphics graphics;
 
-    public EnemyShot(GraphicsManager.Graphics graphics, float originX, float originY) {
+    public PlayerShot(GraphicsManager.Graphics graphics, float originX, float originY) {
         this.graphics = graphics;
         rectangle = new Rectangle();
         rectangle.x = originX;
@@ -22,16 +24,15 @@ public class EnemyShot implements Shot {
         rectangle.width = 10;
     }
 
-
     @Override
     public void render(SpriteBatch batch, float animationTime) {
-        TextureRegion enemyShotFrame = graphics.frameToRender(animationTime);
-        batch.draw(enemyShotFrame, rectangle.x, rectangle.y);
+        TextureRegion shotFrame = graphics.frameToRender(animationTime);
+        batch.draw(shotFrame, rectangle.x, rectangle.y);
     }
 
     @Override
     public void updateState() {
-        this.rectangle.y -= 200 * Gdx.graphics.getDeltaTime();
+        this.rectangle.y += 200 * Gdx.graphics.getDeltaTime();
     }
 
     @Override
@@ -39,7 +40,8 @@ public class EnemyShot implements Shot {
         return targetableObject.checkHit(this.rectangle, targetableObject.rectangle());
     }
 
+    @Override
     public boolean isOutsideScreen() {
-        return this.rectangle.y + 10 < 0;
+        return this.rectangle.y - 10 > HEIGHT;
     }
 }
